redactor/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
httpx>=0.27.0
Pillow>=10.2.0
pypdf2>=3.0.1
natasha>=1.6.0
python-multipart>=0.0.9
pydantic-settings>=2.2.1
python-dotenv>=1.0.1
